requirements.txt
Django>=5.0
djangorestframework>=3.14
psycopg2>=2.9.9
drf-spectacular>=0.26.5
django-cors-headers==4.3.1
django-filter>=23.5
djangorestframework-simplejwt==5.3.1
